rag-backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
python-multipart==0.0.9
tqdm==4.66.4
pypdf==4.2.0
markdown-it-py==3.0.0
beautifulsoup4==4.12.3
python-magic==0.4.27

# Haystack 2 + integrations
haystack-ai>=2.12,<3
qdrant-client>=1.9.0
qdrant-haystack>=9.2.0
ollama-haystack>=4.0.0
